+++ fuzz/fuzz_targets/cmpct_differential.rs
@@ -10,7 +10,8 @@ use libfuzzer_sys::fuzz_target;
 use rbitcoin_fuzz::{spawn_bitcoind_p2p, tmp_dir, CoreChild};
 use rbitcoin_net::{
     classify_v2_cmpct_peer, cmpct_missing_empty_mempool, encode_cmpctblock_v2, encode_pong_v2,
-    encode_sendcmpct_hb_v2, prepare_cmpct_fuzz_hsi, CmpctPeerFrame, NetError, V2PlainSession,
+    encode_sendcmpct_hb_v2, prepare_cmpct_fuzz_hsi, BlockOracle, CmpctPeerFrame, NetError,
+    V2PlainSession,
 };
 use tokio::net::TcpStream;
 use tokio::runtime::{Builder, Runtime};